requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
pandas==1.5.3
numpy==1.24.2
langchain
python-dotenv
requests
plotly